refactor(graphs): rename chart-constants to constants and update imports
Simplify naming convention by renaming chart-constants.ts to constants.ts
and update all import references across graph components, hooks, and utilities.
