+++ ui/components/graphs/ScatterPlot.tsx
@@ -12,7 +12,7 @@ import {
 } from "recharts";
 
 import { AlertPill } from "./shared/AlertPill";
-import { CHART_COLORS } from "./shared/chart-constants";
+import { CHART_COLORS } from "./shared/constants";
 import { ChartLegend } from "./shared/ChartLegend";
 import { getSeverityColorByRiskScore } from "./shared/severity-utils";
