+++ ui/components/graphs/shared/constants.ts
